Fix Playlist rotation
When we use extra html tags in <li> playlist, it does not get the correct selectedIndex and returns -1.
Which cause player to start again from the beginning (0 index after -1)
As you can see in Bar-UI Example, Player 1 and first track. It has more html tags other than single <a> tag. When you play it, it will be played again automatically.
This change in line number 446 fix the issue :)